Refactor: Update package structure and state management
This commit includes several refactoring changes to the package structure and state management within the application.

Key changes:

-   Moved `HomeNavigation` to `home/navigation/HomeNavigation.kt`.
-   Renamed `SignUpScreen` to `MainSignUpScreen`.
-   Deleted `SignInState.kt` and `SignUpState.kt`, replaced with model files in the specific package.
-   Created `SignInState.kt` and `SignUpState.kt` in `model` package.
-   Updated imports to reflect the new package structure.
-   Modified `HomeTopAppBar` to set tint to `Color.Unspecified`.
- Changed `AtSoptSnackbar` text color to `Black`
- Modified `SignUpPasswordScreen` to private function.
- Modified `LocalSnackBarTrigger` comment.
