+++ app/src/main/java/org/sopt/at/core/designsystem/event/SnackbarTrigger.kt
@@ -3,5 +3,5 @@ package org.sopt.at.core.designsystem.event
 import androidx.compose.runtime.staticCompositionLocalOf
 
 val LocalSnackBarTrigger = staticCompositionLocalOf<(String) -> Unit> {
-    error("No SnackBar provided")
+    error("이건 어떤 로직으로 동작할까요?")
 }
