+++ app/src/main/java/org/sopt/at/presentation/home/navigation/HomeNavigation.kt
@@ -1,4 +1,4 @@
-package org.sopt.at.presentation.home
+package org.sopt.at.presentation.home.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -7,6 +7,7 @@ import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import kotlinx.serialization.Serializable
 import org.sopt.at.core.navigation.MainTabRoute
+import org.sopt.at.presentation.home.HomeRoute
 
 fun NavController.navigateToHome(
     navOptions: NavOptions? = null
