+++ app/src/main/java/org/sopt/at/presentation/main/MainScreen.kt
@@ -22,7 +22,7 @@ import kotlinx.coroutines.launch
 import org.sopt.at.core.designsystem.component.snackbar.AtSoptSnackbar
 import org.sopt.at.core.designsystem.event.LocalSnackBarTrigger
 import org.sopt.at.presentation.history.HistoryNavGraph
-import org.sopt.at.presentation.home.homeNavGraph
+import org.sopt.at.presentation.home.navigation.homeNavGraph
 import org.sopt.at.presentation.live.LiveNavGraph
 import org.sopt.at.presentation.main.MainNavigator
 import org.sopt.at.presentation.main.MainTab
