+++ app/src/main/java/org/sopt/at/presentation/signup/MainSignUpScreen.kt
@@ -19,7 +19,7 @@ import org.sopt.at.presentation.signup.navigation.SignUpId
 import org.sopt.at.presentation.signup.navigation.signUpGraph
 
 @Composable
-fun SignUpScreen(
+fun MainSignUpScreen(
     paddingValues: PaddingValues,
     navigateToSingIn: () -> Unit,
     navigateUp: () -> Unit,
